// packages/bespoke/game/Graphical/VickreyAuction3/src/interface.d.ts
export interface ICreateParams {

    /** CreateParams round */
    round?: (number|null);

    /** CreateParams groupSize */
    groupSize?: (number|null);

    /** CreateParams startingPrice */
    startingPrice?: (number|null);

    /** CreateParams privatePrice */
    privatePrice?: (string[]|null);

    /** CreateParams nextPhaseKey */
    nextPhaseKey?: (string|null);
}
// ...
export interface IMoveParams {

    /** MoveParams price */
    price?: (number|null);
}
// ...
export interface IPushParams {

    /** PushParams roundIndex */
    roundIndex?: (number|null);

    /** PushParams dealTimer */
    dealTimer?: (number|null);

    /** PushParams newRoundTimer */
    newRoundTimer?: (number|null);
}
// ...
export interface IGameState {

    /** GameState groups */
    groups?: (GameState.IGroup[]|null);
}

/** Represents a GameState. */
export class GameState implements IGameState {

    /**
     * Constructs a new GameState.
     * @param [properties] Properties to set
     */
    constructor(properties?: IGameState);

    /** GameState groups. */
    public groups: GameState.IGroup[];

    /**
     * Creates a new GameState instance using the specified properties.
     * @param [properties] Properties to set
     * @returns GameState instance
     */
    public static create(properties?: IGameState): GameState;

    /**
     * Encodes the specified GameState message. Does not implicitly {@link GameState.verify|verify} messages.
     * @param message GameState message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encode(message: IGameState, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Encodes the specified GameState message, length delimited. Does not implicitly {@link GameState.verify|verify} messages.
     * @param message GameState message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encodeDelimited(message: IGameState, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Decodes a GameState message from the specified reader or buffer.
     * @param reader Reader or buffer to decode from
     * @param [length] Message length if known beforehand
     * @returns GameState
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): GameState;

    /**
     * Decodes a GameState message from the specified reader or buffer, length delimited.
     * @param reader Reader or buffer to decode from
     * @returns GameState
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decodeDelimited(reader: ($protobuf.Reader|Uint8Array)): GameState;

    /**
     * Verifies a GameState message.
     * @param message Plain object to verify
     * @returns `null` if valid, otherwise the reason why it is not
     */
    public static verify(message: { [k: string]: any }): (string|null);

    /**
     * Creates a GameState message from a plain object. Also converts values to their respective internal types.
     * @param object Plain object
     * @returns GameState
     */
    public static fromObject(object: { [k: string]: any }): GameState;

    /**
     * Creates a plain object from a GameState message. Also converts values to other types if specified.
     * @param message GameState
     * @param [options] Conversion options
     * @returns Plain object
     */
    public static toObject(message: GameState, options?: $protobuf.IConversionOptions): { [k: string]: any };

    /**
     * Converts this GameState to JSON.
     * @returns JSON object
     */
    public toJSON(): { [k: string]: any };
}

export namespace GameState {

    /** Properties of a Group. */
    interface IGroup {

        /** Group roundIndex */
        roundIndex?: (number|null);

        /** Group rounds */
        rounds?: (GameState.Group.IRound[]|null);

        /** Group playerNum */
        playerNum?: (number|null);
    }
// ...
    namespace Group {

        /** Properties of a Round. */
        interface IRound {

            /** Round playerStatus */
            playerStatus?: (number[]|null);
        }
// ...
    }
}

/** Properties of a PlayerState. */
export interface IPlayerState {

    /** PlayerState groupIndex */
    groupIndex?: (number|null);

    /** PlayerState positionIndex */
    positionIndex?: (number|null);

    /** PlayerState privatePrices */
    privatePrices?: (number[]|null);

    /** PlayerState prices */
    prices?: (number[]|null);

    /** PlayerState profits */
    profits?: (number[]|null);
}